requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.3.0
torch>=2.0.0
torchvision>=0.15.0
pytorch3d>=0.7.5
Pillow>=9.0.0
scipy>=1.7.0
imageio>=2.9.0
plotly>=5.0.0
gradio>=3.0.0
